+++ packages/core/src/checkbox/CheckboxIcon.tsx
@@ -4,32 +4,56 @@ import { useWindow } from "@salt-ds/window";
 import { useDensity } from "../salt-provider";
 import { AdornmentValidationStatus } from "../status-adornment";
 import { makePrefixer } from "../utils";
-import {
-  CheckboxCheckedIcon,
-  CheckboxCheckedIconHD,
-  CheckboxIndeterminateIcon,
-  CheckboxUncheckedIcon,
-} from "./assets";
 import checkboxIconCss from "./CheckboxIcon.css";
+import {
+  IconProps,
+  SuccessIcon,
+  SuccessSmallIcon,
+  SuccessSmallSolidIcon,
+  SuccessSolidIcon,
+} from "@salt-ds/icons";
 
 export interface CheckboxIconProps {
   checked?: boolean;
   className?: string;
   disabled?: boolean;
-  error?: boolean /* **Deprecated**: replaced with validationStatus */;
+  /**
+   * @deprecated Use validationStatus instead
+   */
+  error?: boolean;
   indeterminate?: boolean;
+  readOnly?: boolean;
   validationStatus?: AdornmentValidationStatus;
 }
 
 const withBaseName = makePrefixer("saltCheckboxIcon");
 
+function CheckedIcon(props: IconProps) {
+  const density = useDensity();
+  return density === "high" ? (
+    <SuccessSmallSolidIcon {...props} />
+  ) : (
+    <SuccessSolidIcon {...props} />
+  );
+}
+
+function CheckedReadOnlyIcon(props: IconProps) {
+  const density = useDensity();
+  return density === "high" ? (
+    <SuccessSmallIcon {...props} />
+  ) : (
+    <SuccessIcon {...props} />
+  );
+}
+
 export const CheckboxIcon = ({
   checked = false,
-  className: classNameProp,
+  className,
   disabled,
   error,
   indeterminate,
   validationStatus,
+  readOnly,
 }: CheckboxIconProps): JSX.Element => {
   const targetWindow = useWindow();
   useComponentCssInjection({
@@ -37,34 +61,29 @@ export const CheckboxIcon = ({
     css: checkboxIconCss,
     window: targetWindow,
   });
-  const className = clsx(
-    withBaseName(),
-    {
-      [withBaseName("disabled")]: disabled,
-      [withBaseName("error")]: error,
-      [withBaseName(validationStatus || "")]: validationStatus,
-    },
-    classNameProp
-  );
 
-  // A different CheckboxCheckedIcon is rendered if the density is set to high
-  const density = useDensity();
-
-  return indeterminate ? (
-    <CheckboxIndeterminateIcon
-      className={clsx(className, withBaseName("indeterminate"))}
-    />
-  ) : checked ? (
-    density === "high" ? (
-      <CheckboxCheckedIconHD
-        className={clsx(className, withBaseName("checked"))}
-      />
-    ) : (
-      <CheckboxCheckedIcon
-        className={clsx(className, withBaseName("checked"))}
-      />
-    )
-  ) : (
-    <CheckboxUncheckedIcon className={className} />
+  return (
+    <div
+      aria-hidden="true"
+      className={clsx(
+        withBaseName(),
+        {
+          [withBaseName("checked")]: checked,
+          [withBaseName("disabled")]: disabled,
+          [withBaseName("error")]: error,
+          [withBaseName(validationStatus || "")]: validationStatus,
+          [withBaseName("indeterminate")]: indeterminate,
+          [withBaseName("readOnly")]: readOnly,
+        },
+        className
+      )}
+    >
+      {checked && !indeterminate && !readOnly && (
+        <CheckedIcon className={withBaseName("icon")} />
+      )}
+      {checked && !indeterminate && readOnly && (
+        <CheckedReadOnlyIcon className={withBaseName("icon")} />
+      )}
+    </div>
   );
 };