+++ packages/core/stories/form-field/form-field.stories.tsx
@@ -1,4 +1,9 @@
-import { ChangeEvent, ChangeEventHandler, useState } from "react";
+import {
+  ChangeEvent,
+  ChangeEventHandler,
+  CSSProperties,
+  useState,
+} from "react";
 import {
   Checkbox,
   CheckboxGroup,
@@ -19,6 +24,7 @@ import {
 } from "@salt-ds/core";
 import { ComponentMeta, ComponentStory } from "@storybook/react";
 import { NoteIcon, InfoIcon } from "@salt-ds/icons";
+import { MultilineInput } from "@salt-ds/lab";
 
 export default {
   title: "Core/Form Field",
@@ -131,8 +137,19 @@ export const LabelLeft: ComponentStory<typeof FormField> = (props) => {
   );
 };
 
+export const LabelQuestion: ComponentStory<typeof FormField> = (props) => {
+  return (
+    <FormField {...props}>
+      <FormLabel pronounced>
+        Do your current qualifications align with the role? Please describe.
+      </FormLabel>
+      <MultilineInput fullBorder defaultValue="Answer" />
+    </FormField>
+  );
+};
+
 export const MultiChild: ComponentStory<typeof FormField> = (props) => {
-  const [firstValue, setFirstValue] = useState("Abcdef");
+  const [firstValue, setFirstValue] = useState("Five");
   const [secondValue, setSecondValue] = useState("");
 
   const valid = firstValue.length && secondValue.length;
@@ -289,8 +306,8 @@ const radioData = [
     value: "parent",
   },
   {
-    label: "Subsidary",
-    value: "subsidary",
+    label: "Subsidiary",
+    value: "subsidiary",
   },
 ];
 
@@ -400,6 +417,58 @@ export const WithControls: ComponentStory<typeof FormField> = (props) => {
         <FormHelperText>{`${
           isCheckboxError ? "Must select at least one option. " : ""
         }`}</FormHelperText>
+      </FormField>
+    </FlowLayout>
+  );
+};
+
+export const WithMultilineInput: ComponentStory<typeof FormField> = (props) => {
+  return (
+    <FlowLayout style={{ width: "366px" }}>
+      <FormField {...props}>
+        <FormLabel>Form Field label</FormLabel>
+        <MultilineInput defaultValue="Value" />
+        <FormHelperText>Helper text</FormHelperText>
+      </FormField>
+      <FormField readOnly {...props}>
+        <FormLabel>Form Field label</FormLabel>
+        <MultilineInput defaultValue="Value" />
+        <FormHelperText>Helper text</FormHelperText>
+      </FormField>
+      <FormField disabled {...props}>
+        <FormLabel>Form Field label</FormLabel>
+        <MultilineInput defaultValue="Value" />
+        <FormHelperText>Helper text</FormHelperText>
+      </FormField>
+      <FormField validationStatus="error" {...props}>
+        <FormLabel>Form Field label</FormLabel>
+        <MultilineInput defaultValue="Value" />
+        <FormHelperText>Helper text</FormHelperText>
+      </FormField>
+    </FlowLayout>
+  );
+};
+
+export const WithMultilineInputAsQuestion: ComponentStory<typeof FormField> = (
+  props
+) => {
+  return (
+    <FlowLayout style={{ width: "366px" }}>
+      <FormField necessity="required" {...props}>
+        <FormLabel pronounced>
+          Will data related to the new initiative be shared across lines of
+          business, external to the firm, or across-jurisdictions? Please
+          explain.
+        </FormLabel>
+        <MultilineInput defaultValue="Value" />
+      </FormField>
+      <FormField necessity="optional" {...props}>
+        <FormLabel pronounced>
+          For the legal entity country of incorporation, are you aware of any
+          specific regulations relating to oversight of third parties?
+        </FormLabel>
+        <MultilineInput variant="secondary" fullBorder defaultValue="Value" />
+        <FormHelperText>Helper text</FormHelperText>
       </FormField>
     </FlowLayout>
   );
@@ -674,7 +743,7 @@ export const GroupedWithVariant: ComponentStory<typeof FormField> = (props) => {
 export const MultiColumnLayout: ComponentStory<typeof FormField> = (props) => {
   return (
     <StackLayout
-      style={{ "--formField-label-width": "100px" } as React.CSSProperties}
+      style={{ "--formField-label-width": "100px" } as CSSProperties}
     >
       <FormField {...props}>
         <FormLabel>Form Field label left</FormLabel>