+++ tooling/theme-editor/src/tokens/characteristics/CharacteristicPattern.tsx
@@ -2,11 +2,7 @@ import { clsx } from "clsx";
 import { ReactElement } from "react";
 import { useSearchParams } from "react-router-dom";
 import { capitalize, makePrefixer } from "@salt-ds/core";
-import {
-  AccordionDetails,
-  AccordionSection,
-  AccordionSummary,
-} from "@salt-ds/lab";
+import { AccordionPanel, Accordion, AccordionHeader } from "@salt-ds/lab";
 import { JSONObj } from "../../helpers/parseToJson";
 import { ChildrenValuesWithinSection } from "../ChildrenValues";
 import { ScopeLabel } from "../labels/ScopeLabel";
@@ -30,11 +26,12 @@ export const CharacteristicPattern = (
   const [searchParams, setSearchParams] = useSearchParams();
 
   return (
-    <AccordionSection
+    <Accordion
       expanded={props.expandedCharacteristics.includes(props.patternName)}
-      onChange={(isExpanded) => {
+      value={props.patternName}
+      onToggle={() => {
         let characteristics;
-        if (isExpanded) {
+        if (props.expandedCharacteristics.includes(props.patternName)) {
           const openSections = searchParams.get("open");
           characteristics = props.patternName;
           if (openSections) {
@@ -53,10 +50,10 @@ export const CharacteristicPattern = (
           : setSearchParams({});
       }}
     >
-      <AccordionSummary>
+      <AccordionHeader>
         {capitalize(props.patternName) as string}
-      </AccordionSummary>
-      <AccordionDetails>
+      </AccordionHeader>
+      <AccordionPanel>
         {props.values &&
           Object.keys(props.values).map((scope) => {
             return (
@@ -95,7 +92,7 @@ export const CharacteristicPattern = (
               </div>
             );
           })}
-      </AccordionDetails>
-    </AccordionSection>
+      </AccordionPanel>
+    </Accordion>
   );
 };